refactor(dh/esett): align library structure to feature-examples pattern
