+++ libs/dh/esett/feature-balance-responsible/src/drawer/dh-drawer.component.ts
