+++ libs/dh/esett/feature-outgoing-messages/index.ts
@@ -16,5 +16,5 @@
  * limitations under the License.
  */
 //#endregion
-export { DhOutgoingMessagesComponent as default } from './lib/table.component';
-export { DhOutgoingMessageDetailsComponent } from './lib/details.component';
+export { DhOutgoingMessagesComponent as default } from './src/table.component';
+export { DhOutgoingMessageDetailsComponent } from './src/details.component';
