+++ libs/dh/esett/feature-outgoing-messages/src/details.component.ts
