+++ libs/dh/esett/feature-outgoing-messages/src/types.ts
