+++ libs/dh/esett/shell/index.ts
@@ -16,5 +16,4 @@
  * limitations under the License.
  */
 //#endregion
-export { DhBalanceResponsibleComponent as default } from './lib/dh-balance-responsible.component';
-export { DhBalanceResponsibleDrawerComponent } from './lib/drawer/dh-drawer.component';
+export { dhESettShellRoutes as default } from './src/dh-esett-shell.routes';
