+++ libs/dh/esett/shell/tsconfig.lib.json
@@ -7,6 +7,6 @@
     "inlineSources": true,
     "types": []
   },
-  "exclude": ["src/**/*.spec.ts", "src/test-setup.ts", "vite.config.mts", "src/**/*.test.ts"],
+  "exclude": ["tests/**/*.spec.ts", "tests/test-setup.ts", "vite.config.mts", "tests/**/*.test.ts"],
   "include": ["src/**/*.ts"]
 }
