+++ libs/dh/esett/shell/vite.config.mts
@@ -1,25 +1,56 @@
+//#region License
+/**
+ * @license
+ * Copyright 2020 Energinet DataHub A/S
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License2");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+//#endregion
+/// <reference types='vitest' />
 /// <reference types="vitest/config" />
 import { defineConfig } from 'vite';
 import angular from '@analogjs/vite-plugin-angular';
-import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
 
-export default defineConfig({
+import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
+import { nxCopyAssetsPlugin } from '@nx/vite/plugins/nx-copy-assets.plugin';
+
+export default defineConfig(() => ({
   root: __dirname,
-  plugins: [angular({ tsconfig: './tsconfig.json' }), nxViteTsPaths()],
+  cacheDir: '../../../../node_modules/.vite/libs/dh/esett/shell',
+  plugins: [
+    angular({ tsconfig: './tsconfig.json' }),
+    nxViteTsPaths(),
+    nxCopyAssetsPlugin(['*.md']),
+  ],
   test: {
     passWithNoTests: true,
+    watch: false,
     globals: true,
-    environment: 'happy-dom',
-    setupFiles: './src/test-setup.ts',
+    environment: 'jsdom',
     include: ['src/**/*.spec.ts', 'tests/**/*.spec.ts'],
+    setupFiles: ['tests/test-setup.ts'],
+    reporters: ['default'],
     coverage: {
-      enabled: true,
-      provider: 'v8',
-      reporter: ['html', 'json', 'text-summary'],
-      reportsDirectory: '../../../../coverage/shell',
+      reportsDirectory: '../../../../coverage/libs/dh/esett/shell',
+      provider: 'v8' as const,
     },
     pool: 'forks',
     isolate: false,
     maxWorkers: 1,
+    server: {
+      deps: {
+        inline: [/fesm2022/],
+      },
+    },
   },
-});
+}));
